Add fixed_log and fixed_sqrt functions for fixed-point number conversion
Implemented conversion of double data type to unsigned long type.
Added fixed_log function for calculating logarithm of fixed-point numbers.
Added fixed_sqrt function for computing square root of fixed-point numbers.

+++ game.h
@@ -21,5 +21,5 @@ extern const line_t lines[4];
 
 int *available_moves(char *table);
 char check_win(char *t);
-double calculate_win_value(char win, char player);
+unsigned long calculate_win_value(char win, char player);
 void draw_board(const char *t);
